deps/v8/test/inspector/debugger/value-unavailable-scopes-expected.txt

Test scopes with unavailable values

Running test: testTemporalDeadZone
variable y correctly reported as <value_unavailable>

Running test: testOptimizedOut
variable y correctly reported as <value_unavailable>
